+++ container/common/fsHandler.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/google/cadvisor/fs"
 
-	"github.com/golang/glog"
+	"k8s.io/klog"
 )
 
 type FsHandler interface {
@@ -51,7 +51,6 @@ type realFsHandler struct {
 }
 
 const (
-	longOp           = time.Second
 	timeout          = 2 * time.Minute
 	maxBackoffFactor = 20
 )
@@ -93,10 +92,10 @@ func (fh *realFsHandler) update() error {
 	fh.Lock()
 	defer fh.Unlock()
 	fh.lastUpdate = time.Now()
-	if rootDiskErr == nil && fh.rootfs != "" {
+	if rootInodeErr == nil && fh.rootfs != "" {
 		fh.usage.InodeUsage = inodeUsage
 	}
-	if rootInodeErr == nil && fh.rootfs != "" {
+	if rootDiskErr == nil && fh.rootfs != "" {
 		fh.usage.TotalUsageBytes = baseUsage + extraDirUsage
 	}
 	if extraDiskErr == nil && fh.extraDir != "" {
@@ -111,6 +110,7 @@ func (fh *realFsHandler) update() error {
 
 func (fh *realFsHandler) trackUsage() {
 	fh.update()
+	longOp := time.Second
 	for {
 		select {
 		case <-fh.stopChan:
@@ -118,7 +118,7 @@ func (fh *realFsHandler) trackUsage() {
 		case <-time.After(fh.period):
 			start := time.Now()
 			if err := fh.update(); err != nil {
-				glog.Errorf("failed to collect filesystem stats - %v", err)
+				klog.Errorf("failed to collect filesystem stats - %v", err)
 				fh.period = fh.period * 2
 				if fh.period > maxBackoffFactor*fh.minPeriod {
 					fh.period = maxBackoffFactor * fh.minPeriod
@@ -128,7 +128,11 @@ func (fh *realFsHandler) trackUsage() {
 			}
 			duration := time.Since(start)
 			if duration > longOp {
-				glog.V(2).Infof("du and find on following dirs took %v: %v", duration, []string{fh.rootfs, fh.extraDir})
+				// adapt longOp time so that message doesn't continue to print
+				// if the long duration is persistent either because of slow
+				// disk or lots of containers.
+				longOp = longOp + time.Second
+				klog.V(2).Infof("du and find on following dirs took %v: %v; will not log again for this container unless duration exceeds %v", duration, []string{fh.rootfs, fh.extraDir}, longOp)
 			}
 		}
 	}