+++ machine/machine.go
@@ -24,14 +24,15 @@ import (
 
 	// s390/s390x changes
 	"runtime"
-	"syscall"
 
 	info "github.com/google/cadvisor/info/v1"
 	"github.com/google/cadvisor/utils"
 	"github.com/google/cadvisor/utils/sysfs"
 	"github.com/google/cadvisor/utils/sysinfo"
 
-	"github.com/golang/glog"
+	"k8s.io/klog"
+
+	"golang.org/x/sys/unix"
 )
 
 var (
@@ -190,7 +191,7 @@ func GetTopology(sysFs sysfs.SysFs, cpuinfo string) ([]info.Node, int, error) {
 	for idx, node := range nodes {
 		caches, err := sysinfo.GetCacheInfo(sysFs, node.Cores[0].Threads[0])
 		if err != nil {
-			glog.Errorf("failed to get cache information for node %d: %v", node.Id, err)
+			klog.Errorf("failed to get cache information for node %d: %v", node.Id, err)
 			continue
 		}
 		numThreadsPerCore := len(node.Cores[0].Threads)
@@ -265,18 +266,13 @@ func addNode(nodes *[]info.Node, id int) (int, error) {
 
 // s390/s390x changes
 func getMachineArch() (string, error) {
-	uname := syscall.Utsname{}
-	err := syscall.Uname(&uname)
+	uname := unix.Utsname{}
+	err := unix.Uname(&uname)
 	if err != nil {
 		return "", err
 	}
 
-	var arch string
-	for _, val := range uname.Machine {
-		arch += string(int(val))
-	}
-
-	return arch, nil
+	return string(uname.Machine[:]), nil
 }
 
 // arm32 chanes