+++ manager/manager_test.go
@@ -34,7 +34,9 @@ import (
 	itest "github.com/google/cadvisor/info/v1/test"
 	"github.com/google/cadvisor/info/v2"
 	"github.com/google/cadvisor/utils/sysfs/fakesysfs"
+
 	"github.com/stretchr/testify/assert"
+	clock "k8s.io/utils/clock/testing"
 )
 
 // TODO(vmarmol): Refactor these tests.
@@ -59,7 +61,7 @@ func createManagerAndAddContainers(
 			spec,
 			nil,
 		).Once()
-		cont, err := newContainerData(name, memoryCache, mockHandler, false, &collector.GenericCollectorManager{}, 60*time.Second, true)
+		cont, err := newContainerData(name, memoryCache, mockHandler, false, &collector.GenericCollectorManager{}, 60*time.Second, true, clock.NewFakeClock(time.Now()))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -100,8 +102,12 @@ func expectManagerWithContainers(containers []string, query *info.ContainerInfoR
 			if err != nil {
 				t.Error(err)
 			}
+
+			cInfo := info.ContainerInfo{
+				ContainerReference: ref,
+			}
 			for _, stat := range cinfo.Stats {
-				err = memoryCache.AddStats(ref, stat)
+				err = memoryCache.AddStats(&cInfo, stat)
 				if err != nil {
 					t.Error(err)
 				}
@@ -146,8 +152,13 @@ func expectManagerWithContainersV2(containers []string, query *info.ContainerInf
 			if err != nil {
 				t.Error(err)
 			}
+
+			cInfo := info.ContainerInfo{
+				ContainerReference: ref,
+			}
+
 			for _, stat := range cinfo.Stats {
-				err = memoryCache.AddStats(ref, stat)
+				err = memoryCache.AddStats(&cInfo, stat)
 				if err != nil {
 					t.Error(err)
 				}
@@ -261,7 +272,7 @@ func TestGetContainerInfoV2Failure(t *testing.T) {
 	handlerMap[failing].GetSpec() // Use up default GetSpec call, and replace below
 	handlerMap[failing].On("GetSpec").Return(info.ContainerSpec{}, mockErr)
 	handlerMap[failing].On("Exists").Return(true)
-	m.containers[namespacedContainerName{Name: failing}].lastUpdatedTime = time.Time{} // Force GetSpec.
+	m.containers[namespacedContainerName{Name: failing}].infoLastUpdatedTime = time.Time{} // Force GetSpec.
 
 	infos, err := m.GetContainerInfoV2("/", options)
 	if err == nil {
@@ -356,7 +367,7 @@ func TestDockerContainersInfo(t *testing.T) {
 }
 
 func TestNewNilManager(t *testing.T) {
-	_, err := New(nil, nil, 60*time.Second, true, container.MetricSet{}, http.DefaultClient)
+	_, err := New(nil, nil, 60*time.Second, true, container.MetricSet{}, http.DefaultClient, []string{"/"})
 	if err == nil {
 		t.Fatalf("Expected nil manager to return error")
 	}