+++ CHANGELOG.md
@@ -1,5 +1,52 @@
 # Changelog
 
+### 0.32.0 (2018-11-12)
+- Add container process and file descriptor metrics (disabled by default)
+- Rename `type` label to `failure_type` for prometheus `memory_failures_total` metric
+- Reduce mesos error logging when mesos not present
+
+### 0.31.0 (2018-09-07)
+- Fix NVML initialization race condition
+- Fix brtfs filesystem discovery
+- Fix race condition with AllDockerContainers
+- Don't watch .mount cgroups
+- Reduce lock contention during list containers
+- Don't produce prometheus metrics for ignored metrics
+- Add option to not export container labels as prometheus labels
+- Docs: Publish cAdvisor daemonset
+- Docs: Add documentation for exported prometheus metrics
+
+### 0.30.1 (2018-06-11)
+- Revert switch from inotify to fsnotify
+
+### 0.30.0 (2018-06-05)
+- Use IONice to reduce IO priority of `du` and `find`
+- BREAKING API CHANGE: ContainerReference no longer contains Labels.  Use ContainerSpec instead.
+- Add schedstat metrics, disabled by default.
+- Fix a bug where cadvisor failed to discover a sub-cgroup that was created soon after the parent cgroup.
+
+### 0.29.0 (2018-02-20)
+- Disable per-cpu metrics by default for scalability
+- Fix disk usage monitoring of overlayFs
+- Retry docker connection on startup timeout
+
+### 0.28.3 (2017-12-7)
+- Add timeout for docker calls
+- Fix prometheus label consistency
+
+### 0.28.2 (2017-11-21)
+- Fix GPU init race condition
+
+### 0.28.1 (2017-11-20)
+- Add containerd support
+- Fix fsnotify regression from 0.28.0
+- Add on demand metrics
+
+### 0.28.0 (2017-11-06)
+- Add container nvidia GPU metrics
+- Expose container memory max_usage_in_bytes
+- Add container memory reservation to prometheus
+
 ### 0.27.1 (2017-09-06)
 - Add CRI-O support